cleanup - moving email to own routers folder
